Split `GraknGraphPropertyIT`
